requirements.txt
python-dotenv
urllib.parse